Included example of Flask to turn lights on and off
Implemented a simple web application that uses http GET to run the python script that turn the lamps in the smart room on or off.

This web application basically consists of two html buttons that when pressed trigger a Jquery Ajax http GET to the python script.
The Script turns the lights on or off and returns a success or error message.