Includeded 2 examples of Neural netowrks using own datasets.
The didactic_ann.py is implemented manually from scratch.
There are still some things to implement in order to optimize it. Such as BIAS node support and others (TODOs:)

The other one is a Keras didactic_ann that teaches how to use keras and pandas. It will alse be further developed.